delete session storage and all the nasty code
